// bun-impl/src/protocols/game-session-ex-service/types.ts
// AUTOGENERATED by scripts/generate-protocols.ts from ddl/GameSessionExService.json. Do not edit.
import * as c from "../../quazal/codec";
import { GameSessionParticipant, GameSessionSearchResult } from "../game-session-service/types";

export interface GameSessionSearchResultEx {
  gameSessionSearchResult: GameSessionSearchResult;
  participants: GameSessionParticipant[];
}
export const GameSessionSearchResultEx: c.Codec<GameSessionSearchResultEx> = c.struct({
  gameSessionSearchResult: GameSessionSearchResult,
  participants: c.list(GameSessionParticipant),
});